move typedIPC methods to new file
